Rename math_func.js to helpers.js
