+++ samples/js/node/lib/helpers.js
@@ -19,24 +19,17 @@ module.exports = {
 
   prepareResnetTensor: function(arr) {
     /*
-        The images have to normalized using
-        mean_values: [123.675,116.28,103.53] and
-        scale_values: [58.395,57.12,57.375]
+        The images have to be normalized using
+        mean_values: [123.675,116.28,103.53] 
+        and scale_values: [58.395,57.12,57.375]
         of the channels
         For more information refer to:
         https://docs.openvino.ai/latest/ovms_demo_using_onnx_model.html
-        */
-    // [0.485, 0.456, 0.406]
-    // [0.229, 0.224, 0.225]
+    */
     if (arr.length === 0 && arr.length % 3 === 0) {
       return -1;
     }
 
-    // for (var i = 0; i< arr.length; i+=3){
-    //     arr[i] = (arr[i]/255 - 0.485) / 0.229;
-    //     arr[i+1] = (arr[i]/255 - 0.456) / 0.224;
-    //     arr[i+2] = (arr[i]/255 - 0.406) / 0.225;
-    // }
     for (let i = 0; i< arr.length; i+=3) {
       arr[i] = (arr[i] - 103.53) / 57.375;
       arr[i+1] = (arr[i] - 116.28) / 57.12;
