+++ samples/js/node/ppp_example.js
@@ -1,7 +1,7 @@
 var { addon } = require('openvinojs-node');
 
 
-const math = require('./lib/math_func.js');
+const math = require('./lib/helpers.js');
 const Jimp = require('jimp');
 const fs = require('fs');
