+++ samples/js/node/read_model_async_example.js
@@ -2,7 +2,7 @@ var { addon } = require('openvinojs-node');
 
 
 const fs = require('fs');
-const math = require('./lib/math_func.js');
+const math = require('./lib/helpers.js');
 const imagenetClassesMap = require('../assets/imagenet_classes_map.json');
 const Jimp = require('jimp');
